get_blog_posts/requirements.txt
httpx>=0.24.0
beautifulsoup4>=4.12.0
html2text>=2020.1.16
feedparser>=6.0.10
pyyaml>=6.0
requests>=2.31.0
tenacity>=8.2.0
python-dotenv>=1.0.0

